+++ src/Apollo3zehn.OpenApiClientGenerator/Templates/PythonClientTemplate.py
@@ -3,6 +3,8 @@
 {{#Special_NexusFeatures}}
 import asyncio
 import base64
+from concurrent.futures import ThreadPoolExecutor
+from threading import Lock
 {{/Special_NexusFeatures}}
 import json
 {{#Special_NexusFeatures}}